Finishes staff mock set up
