+++ apps/financial-aid/backend/src/app/modules/index.ts
@@ -2,3 +2,4 @@ export { ApplicationModule } from './application'
 export { MunicipalityModule } from './municipality'
 export { FileModule } from './file'
 export { ApplicationEventModule } from './applicationEvent'
+export { StaffModule } from './staff'
