+++ apps/financial-aid/backend/src/app/modules/staff/dto/index.ts
@@ -1,0 +1,1 @@
+export { StaffQueryInput } from './staff.dto'
