+++ apps/financial-aid/backend/src/app/modules/staff/dto/staff.dto.ts
@@ -1,0 +1,10 @@
+import { IsNotEmpty, IsString } from 'class-validator'
+
+import { ApiProperty } from '@nestjs/swagger'
+
+export class StaffQueryInput {
+  @IsNotEmpty()
+  @IsString()
+  @ApiProperty()
+  readonly id!: string
+}
