+++ apps/financial-aid/backend/src/app/modules/staff/index.ts
@@ -1,0 +1,3 @@
+export { StaffModel } from './models'
+export { StaffService } from './staff.service'
+export { StaffModule } from './staff.module'
