+++ apps/financial-aid/backend/src/app/modules/staff/models/index.ts
@@ -1,0 +1,1 @@
+export * from './staff.model'
