+++ libs/financial-aid/shared/src/lib/index.ts
@@ -3,17 +3,7 @@ export * from './formatters'
 export * from './const'
 export * from './cookies'
 export * from './taxCalculator'
-export type {
-  Application,
-  CreateApplication,
-  Municipality,
-  MunicipalitySettings,
-  NavigationProps,
-  User,
-  KeyMapping,
-  GetSignedUrl,
-  SignedUrl,
-  Staff,
-} from './types'
+export type { KeyMapping, StaffRole, UploadFileType } from './types'
 export * from './utils'
 export * from './enums'
+export * from './interfaces'
